ability to awake the agents with hostname
The server can request information and awake the
agent. So far, the URL were generated in different
places in the code.

with this patch, the server generate an array of URL
directly in agent.class.php

It also push on the top of the stack URL that use
hostname and hostname.domain.

IMO, it would be interesting to add a askRemoteStatus()
and a askRemoteRun() directly in agent.class.php and
just call them in taskjob.class.php