kill 'pull' task inactive for 4 hour
If a task has no new log entry for 4 hour, it's probably because
of an agent issue. We kill the task on the server.

We can hardly go under 4h for the moment because the agent can
be busy downloading a large file or processing a complex
installation task.
We will decrease the value once the Deploy task will be able to
send a log update every X minutes during a large download.