fix search form in Normal Task list display
The search form was empty if you select Task Summary display before going to Normal Task display.
Problem was the 2 manageGetValues method calls which seemed to break the form.

Also to be noted, there are $_GET assignments before those calls which are not well documented and
may break in the future due to PHP or GLPI API. I could not find if there were mandatory to this
function.