fix warning about session regenerated by Session class
When calling auth method in Session class, the session is regenerated
and PHP says headers are already sent but we don't care since this is
used in a CLI script. The fix is to disable cookies on session which is
checked by PHP when using the session_regenerate_id() function.