Adds attempts for proxy-related HTTP headers to determine contact IP address
When a proxy is used to contact GLPI, $_SERVER['REMOTE_ADDR'] may not
contains the agent's IP address. The true IP can be in "X-Forwarded-For" or
"X-Real-IP" HTTP headers (depends on proxy's configuration).

This commits checks for theses 2 headers and use the provided IP address
(if valid) as "Public contact address".